requirements.txt
asgiref==3.5.0
astroid==2.9.3
black==21.12b0
click==8.0.3
cycler==0.11.0
Django==4.0.1
djangorestframework==3.13.1
fonttools==4.29.0
isort==5.10.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
matplotlib==3.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
wrapt==1.13.3
